refactor for maintainability and support script injection via manifest.json
