+++ ui/.build/src/build.ts
@@ -1,29 +1,31 @@
 import fs from 'node:fs';
 import cps from 'node:child_process';
 import ps from 'node:process';
-import { parsePackages } from './parse.ts';
-import { tsc, stopTsc } from './tsc.ts';
+import path from 'node:path';
+import { parsePackages, globArray } from './parse.ts';
+import { tsc, stopTscWatch } from './tsc.ts';
 import { sass, stopSass } from './sass.ts';
-import { esbuild, stopEsbuild } from './esbuild.ts';
+import { esbuild, stopEsbuildWatch } from './esbuild.ts';
 import { sync, stopSync } from './sync.ts';
-import { monitor, stopMonitor } from './monitor.ts';
-import { writeManifest } from './manifest.ts';
-import { type Package, env, errorMark, colors as c } from './main.ts';
-import { i18n, stopI18n } from './i18n.ts';
+import { stopManifest } from './manifest.ts';
+import { env, errorMark, colors as c } from './env.ts';
+import { i18n, stopI18nWatch } from './i18n.ts';
+import { unique } from './algo.ts';
+import { clean } from './clean.ts';
 
 export async function build(pkgs: string[]): Promise<void> {
-  if (env.install)
-    cps.execSync('pnpm install --prefer-frozen-lockfile', { cwd: env.rootDir, stdio: 'inherit' });
+  if (env.install) cps.execSync('pnpm install', { cwd: env.rootDir, stdio: 'inherit' });
   if (!pkgs.length) env.log(`Parsing packages in '${c.cyan(env.uiDir)}'`);
 
   ps.chdir(env.uiDir);
-  [env.packages, env.deps] = await parsePackages();
+  await parsePackages();
 
   pkgs
     .filter(x => !env.packages.has(x))
     .forEach(x => env.exit(`${errorMark} - unknown package '${c.magenta(x)}'`));
 
-  env.building = pkgs.length === 0 ? [...env.packages.values()] : depsMany(pkgs);
+  env.building =
+    pkgs.length === 0 ? [...env.packages.values()] : unique(pkgs.flatMap(p => env.transitiveDeps(p)));
 
   if (pkgs.length) env.log(`Building ${c.grey(env.building.map(x => x.name).join(', '))}`);
 
@@ -36,43 +38,60 @@ export async function build(pkgs: string[]): Promise<void> {
   ]);
 
   await Promise.all([sass(), sync(), i18n()]);
-  await tsc().then(esbuild);
-  monitor(pkgs);
+  await Promise.all([tsc(), esbuild()]);
+  if (env.watch) monitor(pkgs);
 }
 
-export async function stopBuild(): Promise<void> {
-  stopMonitor();
+export async function stopBuildWatch(): Promise<void> {
+  for (const w of watchers) w.close();
+  watchers.length = 0;
+  clearTimeout(tscTimeout);
+  clearTimeout(packageTimeout);
+  tscTimeout = packageTimeout = undefined;
   stopSass();
-  stopTsc();
   stopSync();
-  stopI18n();
-  await stopEsbuild();
+  stopI18nWatch();
+  stopManifest();
+  await Promise.allSettled([stopTscWatch(), stopEsbuildWatch()]);
 }
 
-export function postBuild(): void {
-  writeManifest();
+const watchers: fs.FSWatcher[] = [];
+
+let packageTimeout: NodeJS.Timeout | undefined;
+let tscTimeout: NodeJS.Timeout | undefined;
+
+async function monitor(pkgs: string[]): Promise<void> {
+  const [typePkgs, typings] = await Promise.all([
+    globArray('*/package.json', { cwd: env.typesDir }),
+    globArray('*/*.d.ts', { cwd: env.typesDir }),
+  ]);
+  const tscChange = async () => {
+    if (packageTimeout) return;
+    stopManifest();
+    await Promise.allSettled([stopTscWatch(), stopEsbuildWatch()]);
+    clearTimeout(tscTimeout);
+    tscTimeout = setTimeout(() => {
+      if (packageTimeout) return;
+      tsc().then(esbuild);
+    }, 2000);
+  };
+  const packageChange = async () => {
+    if (env.watch && env.install) {
+      clearTimeout(tscTimeout);
+      clearTimeout(packageTimeout);
+      await stopBuildWatch();
+      packageTimeout = setTimeout(() => clean().then(() => build(pkgs)), 2000);
+      return;
+    }
+    env.warn('Exiting due to package.json change');
+    ps.exit(0);
+  };
+
+  watchers.push(fs.watch(path.join(env.rootDir, 'package.json'), packageChange));
+  for (const p of typePkgs) watchers.push(fs.watch(p, packageChange));
+  for (const t of typings) watchers.push(fs.watch(t, tscChange));
   for (const pkg of env.building) {
-    pkg.post.forEach((args: string[]) => {
-      env.log(`[${c.grey(pkg.name)}] exec - ${c.cyanBold(args.join(' '))}`);
-      const stdout = cps.execSync(`${args.join(' ')}`, { cwd: pkg.root });
-      if (stdout) env.log(stdout, { ctx: pkg.name });
-    });
+    watchers.push(fs.watch(path.join(pkg.root, 'package.json'), packageChange));
+    watchers.push(fs.watch(path.join(pkg.root, 'tsconfig.json'), tscChange));
   }
 }
-
-export function prePackage(pkg: Package | undefined): void {
-  pkg?.pre.forEach((args: string[]) => {
-    env.log(`[${c.grey(pkg.name)}] exec - ${c.cyanBold(args.join(' '))}`);
-    const stdout = cps.execSync(`${args.join(' ')}`, { cwd: pkg.root });
-    if (stdout) env.log(stdout, { ctx: pkg.name });
-  });
-}
-
-function depsOne(pkgName: string): Package[] {
-  const collect = (dep: string): string[] => [...(env.deps.get(dep) || []).flatMap(d => collect(d)), dep];
-  return unique(collect(pkgName).map(name => env.packages.get(name)));
-}
-
-const depsMany = (pkgNames: string[]): Package[] => unique(pkgNames.flatMap(depsOne));
-
-const unique = <T>(pkgs: (T | undefined)[]): T[] => [...new Set(pkgs.filter(x => x))] as T[];
