+++ ui/.build/src/esbuild.ts
@@ -1,27 +1,21 @@
 import path from 'node:path';
 import es from 'esbuild';
-import { prePackage } from './build.ts';
-import { env, errorMark, colors as c } from './main.ts';
-import { jsManifest } from './manifest.ts';
+import fs from 'node:fs';
+import { env, errorMark, colors as c } from './env.ts';
+import { type Manifest, updateManifest } from './manifest.ts';
+import { readable } from './parse.ts';
 
-const bundles = new Map<string, string>();
 const esbuildCtx: es.BuildContext[] = [];
-
-export async function stopEsbuild(): Promise<void> {
-  const proof = Promise.allSettled(esbuildCtx.map(x => x.dispose()));
-  esbuildCtx.length = 0;
-  bundles.clear();
-  await proof;
-}
+const inlineWatch: fs.FSWatcher[] = [];
+let inlineTimer: NodeJS.Timeout;
 
 export async function esbuild(): Promise<void> {
   if (!env.esbuild) return;
 
   const entryPoints = [];
   for (const pkg of env.building) {
-    prePackage(pkg);
-    for (const bundle of pkg.bundles ?? []) {
-      entryPoints.push(path.join(pkg.root, bundle));
+    for (const { module } of pkg.bundle) {
+      if (module) entryPoints.push(path.join(pkg.root, module));
     }
   }
   entryPoints.sort();
@@ -39,7 +33,7 @@ export async function esbuild(): Promise<void> {
     outdir: env.jsOutDir,
     entryNames: '[name].[hash]',
     chunkNames: 'common.[hash]',
-    plugins: [onEndPlugin],
+    plugins,
   });
   if (env.watch) {
     ctx.watch();
@@ -50,26 +44,114 @@ export async function esbuild(): Promise<void> {
   }
 }
 
-const onEndPlugin = {
-  name: 'onEnd',
-  setup(build: es.PluginBuild) {
-    build.onEnd(async (result: es.BuildResult) => {
-      for (const err of result.errors) esbuildMessage(err, true);
-      for (const warn of result.warnings) esbuildMessage(warn);
-      if (result.errors.length === 0) jsManifest(result.metafile!);
-      env.done(result.errors.length, 'esbuild');
-    });
-  },
-};
+export async function stopEsbuildWatch(): Promise<void> {
+  const proof = Promise.allSettled(esbuildCtx.map(x => x.dispose()));
+  for (const w of inlineWatch) w.close();
+  inlineWatch.length = 0;
+  esbuildCtx.length = 0;
+  await proof;
+}
 
-function esbuildMessage(msg: es.Message, error = false) {
-  const file = msg.location?.file.replace(/^[./]*/, '') ?? '<unknown>';
-  const line = msg.location?.line
-    ? `:${msg.location.line}`
-    : '' + (msg.location?.column ? `:${msg.location.column}` : '');
-  const srcText = msg.location?.lineText;
-  env.log(`${error ? errorMark : c.warn('WARNING')} - '${c.cyan(file + line)}' - ${msg.text}`, {
-    ctx: 'esbuild',
-  });
-  if (srcText) env.log('  ' + c.magenta(srcText), { ctx: 'esbuild' });
+const plugins = [
+  {
+    name: 'onBundleDone',
+    setup(build: es.PluginBuild) {
+      build.onEnd(async (result: es.BuildResult) => {
+        esbuildLog(result.errors, true);
+        esbuildLog(result.warnings);
+        env.done(result.errors.length, 'esbuild');
+        if (result.errors.length === 0) jsManifest(result.metafile!);
+      });
+    },
+  },
+];
+
+function esbuildLog(msgs: es.Message[], error = false): void {
+  for (const msg of msgs) {
+    const file = msg.location?.file.replace(/^[./]*/, '') ?? '<unknown>';
+    const line = msg.location?.line
+      ? `:${msg.location.line}`
+      : '' + (msg.location?.column ? `:${msg.location.column}` : '');
+    const srcText = msg.location?.lineText;
+    env.log(`${error ? errorMark : c.warn('WARNING')} - '${c.cyan(file + line)}' - ${msg.text}`, {
+      ctx: 'esbuild',
+    });
+    if (srcText) env.log('  ' + c.magenta(srcText), { ctx: 'esbuild' });
+  }
+}
+
+async function jsManifest(meta: es.Metafile = { inputs: {}, outputs: {} }) {
+  for (const w of inlineWatch) w.close();
+  inlineWatch.length = 0;
+  clearTimeout(inlineTimer);
+
+  const newJsManifest: Manifest = {};
+  for (const [filename, info] of Object.entries(meta.outputs)) {
+    const out = parsePath(filename);
+    if (!out) continue;
+    if (out.name === 'common') {
+      out.name = `common.${out.hash}`;
+      newJsManifest[out.name] = {};
+    } else newJsManifest[out.name] = { hash: out.hash };
+    const imports: string[] = [];
+    for (const imp of info.imports) {
+      if (imp.kind === 'import-statement') {
+        const path = parsePath(imp.path);
+        if (path) imports.push(`${path.name}.${path.hash}.js`);
+      }
+    }
+    newJsManifest[out.name].imports = imports;
+  }
+  await inlineManifest(newJsManifest);
+}
+
+async function inlineManifest(js: Manifest) {
+  const makeWatchers = env.watch && inlineWatch.length === 0;
+  let success = true;
+  for (const pkg of env.building) {
+    for (const bundle of pkg.bundle ?? []) {
+      if (!bundle.inline) continue;
+
+      const inlineSrc = path.join(pkg.root, bundle.inline);
+      const moduleName = bundle.module
+        ? path.basename(bundle.module, '.ts')
+        : path.basename(bundle.inline, '.inline.ts');
+      const packageError = `${errorMark} - Package error ${c.blue(JSON.stringify(bundle))}`;
+
+      if (!(await readable(inlineSrc))) {
+        env.log(packageError);
+        for (const w of inlineWatch) w.close();
+        inlineWatch.length = 0;
+        if (!env.watch) env.exit('Failed'); // all inline sources must exist
+      }
+
+      try {
+        const res = await es.transform(await fs.promises.readFile(inlineSrc), {
+          minify: true,
+          loader: 'ts',
+        });
+        esbuildLog(res.warnings);
+        js[moduleName] ??= {};
+        js[moduleName].inline = res.code;
+      } catch (e) {
+        if (e && typeof e === 'object' && 'errors' in e) esbuildLog((e as es.TransformFailure).errors, true);
+        else env.log(`${packageError} - ${JSON.stringify(e)}`);
+        if (env.watch) success = false;
+        else env.exit('Failed');
+      }
+      if (makeWatchers)
+        inlineWatch.push(
+          fs.watch(inlineSrc, () => {
+            clearTimeout(inlineTimer);
+            inlineTimer = setTimeout(() => inlineManifest(js), 200);
+          }),
+        );
+    }
+  }
+  if (success) updateManifest({ js });
+}
+
+function parsePath(path: string) {
+  const match = path.match(/\/public\/compiled\/(.*)\.([A-Z0-9]+)\.js$/);
+  return match ? { name: match[1], hash: match[2] } : undefined;
 }
