+++ ui/.build/src/tsc.ts
@@ -1,19 +1,14 @@
 import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
+import ts from 'typescript';
 import { Worker } from 'node:worker_threads';
-import { env, colors as c, errorMark } from './main.ts';
-import { globArray, folderSize } from './parse.ts';
+import { env, colors as c, errorMark } from './env.ts';
+import { globArray, folderSize, readable } from './parse.ts';
 import { clamp } from './algo.ts';
 import type { WorkerData, Message } from './tscWorker.ts';
-import ts from 'typescript';
 
 const workers: Worker[] = [];
-
-export async function stopTsc(): Promise<void> {
-  await Promise.allSettled(workers.map(w => w.terminate()));
-  workers.length = 0;
-}
 
 export async function tsc(): Promise<void> {
   if (!env.tsc) return;
@@ -42,11 +37,16 @@ export async function tsc(): Promise<void> {
       )
       .push(cfg);
 
-  tscLog(`Transpiling ${c.grey('noCheck')} (${workBuckets.noCheck.length} workers)`);
+  tscLog(`Typing ${c.grey('noCheck')} (${workBuckets.noCheck.length} workers)`);
   await assignWork(workBuckets.noCheck, 'noCheck');
 
   tscLog(`Typechecking ${c.grey('noEmit')} (${workBuckets.noEmit.length} workers)`);
   await assignWork(workBuckets.noEmit, 'noEmit');
+}
+
+export async function stopTscWatch(): Promise<void> {
+  await Promise.allSettled(workers.map(w => w.terminate()));
+  workers.length = 0;
 }
 
 function assignWork(buckets: SplitConfig[][], key: 'noCheck' | 'noEmit') {
@@ -83,9 +83,9 @@ function assignWork(buckets: SplitConfig[][], key: 'noCheck' | 'noEmit') {
 }
 
 // the splitConfig transform generates noCheck and noEmit tsconfigs within the 'build' temp folder.
-// each package that emits gets a --noCheck pass for fast transpilations and declarations.
-// then we do a --noEmit pass on EVERY package to verify things, regardless of its original emit.
-// the declarations from the --noCheck pass allow efficient parallel type checking.
+// each workspace depndency package gets a --noCheck pass for fast declaration generation.
+// then we do a --noEmit pass on EVERY package to verify things. the declarations from the --noCheck
+// pass allow more efficient parallel type checking on most architectures
 
 interface SplitConfig {
   type: 'noEmit' | 'noCheck';
@@ -121,7 +121,7 @@ async function splitConfig(cfgPath: string): Promise<SplitConfig[]> {
     ?.filter((glob: string) => !env.test || !glob.includes('tests'))
     .map((glob: string) => path.resolve(root, glob.replace('${configDir}', '.')));
   config.extends = undefined;
-  config.references = env.deps
+  config.references = env.workspaceDeps
     .get(pkgName)
     ?.map(ref => ({ path: path.join(env.buildTempDir, 'noCheck', `${ref}.tsconfig.json`) }));
 
@@ -133,7 +133,7 @@ async function splitConfig(cfgPath: string): Promise<SplitConfig[]> {
     'noEmit',
     `${pkgName}.tsbuildinfo`,
   );
-  if (env.test && fs.existsSync(path.join(root, 'tests'))) {
+  if (env.test && (await readable(path.join(root, 'tests')))) {
     noEmitData.include.push(path.join(root, 'tests'));
     noEmitData.compilerOptions.rootDir = root;
     noEmitData.compilerOptions.skipLibCheck = true;
@@ -147,6 +147,7 @@ async function splitConfig(cfgPath: string): Promise<SplitConfig[]> {
     const noCheckData = structuredClone(config);
     const noCheck = path.join(env.buildTempDir, 'noCheck', `${pkgName}.tsconfig.json`);
     noCheckData.compilerOptions.noCheck = true;
+    noCheckData.compilerOptions.emitDeclarationOnly = true;
     noCheckData.compilerOptions.tsBuildInfoFile = path.join(
       env.buildTempDir,
       'noCheck',
