+++ ui/.build/usage.txt
@@ -1,10 +1,5 @@
 Usage:
-  ui/build <options> <packages>  # <packages> are top level directories in ui
-                                 # if no packages are specified, all will be processed
-                                 # multiple short options can be preceded by a single dash
-
-Recommended:
-  ui/build -cdw  # clean, build debug, and watch for changes with clean rebuilds
+  ui/build <options>  # multiple short options can be preceded by a single dash
 
 Options:
   -h, --help          show this help and exit
@@ -22,7 +17,7 @@ Options:
   --no-time           don't log the time
   --no-context        don't log the context
 
-Exclusive Options: (any of these will disable other functions)
+Exclusive Options:    (any of these will disable other functions)
   --clean-exit        clean all build artifacts and exit
   --tsc               run tsc on {package}/tsconfig.json and dependencies
   --sass              run sass on {package}/css/build/*.scss and dependencies
@@ -30,10 +25,12 @@ Exclusive Options: (any of these will disable other functions)
   --sync              run sync copies (given in {package}/package.json/lichess/sync objects)
   --i18n              build @types/lichess/i18n.d.ts and translation/js files
 
-Examples:
-  ./build -np              # no pnpm install, build minified
-  ./build -w dasher chart  # watch mode for dasher and chart packages
-  ./build --tsc -w         # watch mode but type checking only
-  ./build -w -l=/path      # build, watch. patch console methods in emitted js to also POST messages
-                           # to ${location.origin}/path. ui/build watch process displays received
-                           # messages as 'web' in build logs
+Recommended:
+  ui/build -cdw       # clean, build debug, and watch for changes with clean rebuilds
+
+Other Examples:
+  ./build -np         # no pnpm install, build minified
+  ./build --tsc -w    # watch mode but type checking only
+  ./build -dwl=/x     # build debug, watch. patch console methods in emitted js to POST log statements
+                        to ${location.origin}/x. ui/build watch process displays messages received
+                        via http/s on this endpoint as 'web' in build logs
